refactor(hir): isolate legacy builders under compat
