+++ test/_fixtures/scenario-corpus.ts
@@ -10,8 +10,8 @@ import type {
 } from "@css-module-explainer/shared";
 import { detectClassUtilImports, scanCxImports } from "../../server/src/core/cx/binding-detector";
 import { parseClassRefs } from "../../server/src/core/cx/class-ref-parser";
-import { buildSourceDocumentFromLegacy } from "../../server/src/core/hir/builders/ts-source-adapter";
-import { buildStyleDocumentFromClassMap } from "../../server/src/core/hir/builders/style-adapter";
+import { buildSourceDocumentFromLegacy } from "../../server/src/core/hir/compat/source-document-builder-compat";
+import { buildStyleDocumentFromClassMap } from "../../server/src/core/hir/compat/style-document-builder-compat";
 import { sourceDocumentToLegacyClassRefs } from "../../server/src/core/hir/compat/source-document-compat";
 import { styleDocumentToLegacyClassMap } from "../../server/src/core/hir/compat/style-document-compat";
 import type { SourceDocumentHIR } from "../../server/src/core/hir/source-types";
