+++ test/_fixtures/test-helpers.ts
@@ -6,8 +6,8 @@ import type {
   SelectorInfo,
   StyleImport,
 } from "@css-module-explainer/shared";
-import { buildStyleDocumentFromClassMap } from "../../server/src/core/hir/builders/style-adapter";
-import { buildSourceDocumentFromLegacy } from "../../server/src/core/hir/builders/ts-source-adapter";
+import { buildSourceDocumentFromLegacy } from "../../server/src/core/hir/compat/source-document-builder-compat";
+import { buildStyleDocumentFromClassMap } from "../../server/src/core/hir/compat/style-document-builder-compat";
 import { SourceFileCache } from "../../server/src/core/ts/source-file-cache";
 import { DocumentAnalysisCache } from "../../server/src/core/indexing/document-analysis-cache";
 import { NullSemanticWorkspaceReferenceIndex } from "../../server/src/core/semantic/workspace-reference-index";
