+++ test/unit/hir/source-document-compat.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest";
 import type { ClassRef, StyleImport } from "@css-module-explainer/shared";
-import { buildSourceDocumentFromLegacy } from "../../../server/src/core/hir/builders/ts-source-adapter";
+import { buildSourceDocumentFromLegacy } from "../../../server/src/core/hir/compat/source-document-builder-compat";
 import { sourceDocumentToLegacyClassRefs } from "../../../server/src/core/hir/compat/source-document-compat";
 
 const ZERO = { start: { line: 0, character: 0 }, end: { line: 0, character: 1 } };
